Added differing fulltext test
