+++ adsft/tests/test_differing_full_text.py
@@ -9,7 +9,7 @@ from datetime import datetime
 import json
 from mock import patch
 
-class TestForcedExtractor(test_base.TestGeneric):
+class TestNonExtractedExtraction(test_base.TestGeneric):
     """
     Class for testing that a file is force extracted if specified by the user.
     """
@@ -21,7 +21,7 @@ class TestForcedExtractor(test_base.TestGeneric):
         worker as well into a class attribute so it is easier to access.
         :return:
         """
-        super(TestForcedExtractor, self).setUp()
+        super(TestNonExtractedExtraction, self).setUp()
         #self.dict_item = {'ft_source': self.test_stub_xml,
                           #'file_format': 'xml',
                           #'provider': 'MNRAS'}
@@ -41,9 +41,9 @@ class TestForcedExtractor(test_base.TestGeneric):
         """
 
         self.clean_up_path(self.expected_paths)
-        super(TestForcedExtractor, self).tearDown()
+        super(TestNonExtractedExtraction, self).tearDown()
 
-    def test_forced_extraction(self):
+    def test_extraction_of_non_extracted(self):
         """
         Tests that when a user specifies 'force_extract' that the full text
         is extracted regardless of its underlying reason for being or not
@@ -56,7 +56,7 @@ class TestForcedExtractor(test_base.TestGeneric):
 
         # User loads the list of full text files and publishes them to the
         # first queue
-        records = read_links_from_file(self.test_publish, force_extract=True)
+        records = read_links_from_file(self.test_publish, force_extract=False)
 
         self.helper_get_details(self.test_publish)
         self.assertEqual(
@@ -67,12 +67,25 @@ class TestForcedExtractor(test_base.TestGeneric):
 
         self.assertTrue(len(records.payload) == 1)
 
+        # Make the fake data to use
+        if not os.path.exists(self.meta_path):
+            os.makedirs(self.meta_path)
+
+        test_meta_content = {
+            'index_date': datetime.utcnow().isoformat()+'Z',
+            'bibcode': 'test4',
+            'provider': 'mnras',
+            'ft_source': 'wrong_source'
+        }
+        with open(self.test_expected, 'w') as test_meta_file:
+            json.dump(test_meta_content, test_meta_file)
+
         # Call the task to check if it should be extracted but mock the extraction task
         with patch.object(tasks.task_extract, 'delay', return_value=None) as task_extract:
             message = records.payload[0]
             tasks.task_check_if_extract(message)
             self.assertTrue(task_extract.called)
-            expected = {'UPDATE': 'FORCE_TO_EXTRACT',
+            expected = {'UPDATE': 'DIFFERING_FULL_TEXT',
                          'bibcode': 'test4',
                          'file_format': 'txt',
                          'ft_source': '/home/docker/workspace/remote/vhost/Sync/Data/Boston/Code/ADSfulltext/tests/test_unit/stub_data/test.txt',
@@ -100,7 +113,7 @@ class TestForcedExtractor(test_base.TestGeneric):
                 with open(meta_path, 'r') as meta_file:
                     meta_content = meta_file.read()
                 self.assertTrue(
-                    'FORCE_TO_EXTRACT' in meta_content,
+                    'DIFFERING_FULL_TEXT' in meta_content,
                     'meta file does not contain the right extract keyword: {0}'
                     .format(meta_content)
                 )
